refactor: use blueprint to classify APIs
