+++ static/css/components/login.css
@@ -97,6 +97,6 @@ input{
     font-weight: 500;
 }
 .loginSuccessfulMessage{
-    color: rgb(174, 240, 188);
+    color: #1da56c;
     font-weight: 500;
 }
